imports: remove edges spack.container.*
